Extract TextEditorMenu from TextEditor
RISDEV-4432
